+++ src/main/kotlin/app/revanced/patches/youtube/layout/autoplaybutton/patch/HideAutoplayButton.kt
@@ -18,9 +18,9 @@ import app.revanced.patches.youtube.misc.mapping.patch.ResourceIdMappingProvider
 import org.jf.dexlib2.iface.instruction.WideLiteralInstruction
 
 @Patch
-@Dependencies(dependencies = [ResourceIdMappingProviderResourcePatch::class])
+@Dependencies([ResourceIdMappingProviderResourcePatch::class])
 @Name("hide-autoplay-button")
-@Description("Disables the autoplay button.")
+@Description("Hides the autoplay button in the video player.")
 @AutoplayButtonCompatibility
 @Version("0.0.1")
 class HideAutoplayButton : BytecodePatch(