+++ src/main/kotlin/app/revanced/patches/youtube/layout/branding/icon/patch/CustomBrandingPatch.kt
@@ -15,11 +15,9 @@ import app.revanced.patches.youtube.misc.manifest.patch.FixLocaleConfigErrorPatc
 import java.nio.file.Files
 
 @Patch
-@Dependencies(
-    dependencies = [FixLocaleConfigErrorPatch::class]
-)
+@Dependencies([FixLocaleConfigErrorPatch::class])
 @Name("custom-branding")
-@Description("Changes the branding of YouTube.")
+@Description("Changes the YouTube launcher icon to be ReVanced's.")
 @CustomBrandingCompatibility
 @Version("0.0.1")
 class CustomBrandingPatch : ResourcePatch() {