+++ src/main/kotlin/app/revanced/patches/youtube/layout/createbutton/patch/CreateButtonRemoverPatch.kt
@@ -22,9 +22,9 @@ import org.jf.dexlib2.iface.instruction.WideLiteralInstruction
 import org.jf.dexlib2.iface.reference.MethodReference
 
 @Patch
-@Dependencies(dependencies = [IntegrationsPatch::class, ResourceIdMappingProviderResourcePatch::class])
+@Dependencies([IntegrationsPatch::class, ResourceIdMappingProviderResourcePatch::class])
 @Name("disable-create-button")
-@Description("Disables the create button.")
+@Description("Hides the create button in the navigation bar.")
 @CreateButtonCompatibility
 @Version("0.0.1")
 class CreateButtonRemoverPatch : BytecodePatch(