+++ src/main/kotlin/app/revanced/patches/youtube/layout/shorts/button/patch/ShortsButtonRemoverPatch.kt
@@ -17,9 +17,9 @@ import app.revanced.patches.youtube.misc.integrations.patch.IntegrationsPatch
 import org.jf.dexlib2.iface.instruction.OneRegisterInstruction
 
 @Patch
-@Dependencies(dependencies = [IntegrationsPatch::class])
+@Dependencies([IntegrationsPatch::class])
 @Name("hide-shorts-button")
-@Description("Hides the shorts button.")
+@Description("Hides the shorts button on the navigation bar.")
 @ShortsButtonCompatibility
 @Version("0.0.1")
 class ShortsButtonRemoverPatch : BytecodePatch(