+++ src/main/kotlin/app/revanced/patches/youtube/layout/watermark/patch/HideWatermarkPatch.kt
@@ -19,9 +19,9 @@ import app.revanced.patches.youtube.layout.watermark.fingerprints.HideWatermarkF
 import app.revanced.patches.youtube.misc.integrations.patch.IntegrationsPatch
 
 @Patch
-@Dependencies(dependencies = [IntegrationsPatch::class])
+@Dependencies([IntegrationsPatch::class])
 @Name("hide-watermark")
-@Description("Hides the creator's watermark on videos.")
+@Description("Hides creator's watermarks on videos.")
 @HideWatermarkCompatibility
 @Version("0.0.1")
 class HideWatermarkPatch : BytecodePatch(