+++ src/main/kotlin/app/revanced/patches/youtube/misc/customplaybackspeed/patch/CustomPlaybackSpeedPatch.kt
@@ -24,8 +24,8 @@ import org.jf.dexlib2.iface.reference.MethodReference
 
 @Patch
 @Name("custom-playback-speed")
-@Description("Allows to change the default playback speed options.")
-@Dependencies(dependencies = [IntegrationsPatch::class])
+@Description("Adds more video playback speed options.")
+@Dependencies([IntegrationsPatch::class])
 @CustomPlaybackSpeedCompatibility
 @Version("0.0.1")
 class CustomPlaybackSpeedPatch : BytecodePatch(